separate dashboard and list user and added drawer to dashboard
